app/material_review/tencent_requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
Pillow
PyMuPDF
tencentcloud-sdk-python